Sortable plugin now wraps the result of headerFormatter
This brings the formatter inline with user expectations. The formatter
is always called with the raw header value not a formatted block of
html from the plug-in. This keeps rules about escaping content clear,
if the user provides a formatter it is their responsibility to escape
html otherwise the library will do it.

Another benefit of this approach is that hacks/html parsing is no
longer required to add extra content inside the sortable anchor.